user setting done
